working on partitioning (not running for large networks)
