+++ backend/fpgaconvnet/parser.py
@@ -1,6 +1,8 @@
 import fpgaconvnet_optimiser.tools.parser as parser
 from .node import FPGAConvNetWrapper
 from optimiser import Network
+from optimiser import Partition
+import copy
 
 def parse(filepath):
 
@@ -8,10 +10,18 @@ def parse(filepath):
     _, graph = parser.parse_net(filepath)
 
     # convert into the node wrappers
-    network = Network()
+    reference = Partition()
+    prev_node = None
     for node in graph.nodes:
         new_node = FPGAConvNetWrapper(graph.nodes[node]["hw"])
-        network.add_node(node, hw=new_node)
+        reference.add_node(node, hw=new_node)
+        # add edge to graph
+        if prev_node != None:
+            reference.add_edge(prev_node, node)
+        prev_node = node
+
+    # create network from reference design
+    network = Network(reference)
 
     # return the wrapped network
     return network
