+++ optimiser/__init__.py
@@ -1,3 +1,4 @@
 from .node import Node
 from .network import Network
+from .partition import Partition
 from .brute import BruteForce
